third_party/WebKit/LayoutTests/media/video-set-rate-from-pause.html

<!DOCTYPE html>
<title>Test that setting a non-zero rate causes an async timeupdate event.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.onplay = t.step_func(function() {
        video.ontimeupdate = t.step_func_done();
        video.playbackRate = 1;
    });

    video.src = "content/test.ogv";
    video.playbackRate = 0;
    video.play();
});
</script>
